Fix context-already-closed error and exception test flake
The parent-child context hook was not creating a new ref for child context reactivations, and then trying to re-close
the old one. In particular this happens during the beginning of an invocation with arguments as EAV tuple nodes for
the arguments are activated before the invocation entry point.

The exception test was flaking due to some runs taking longer than the emission monitor timeout.