Fix several memory leaks
This fixes the memory leaks surrounding contexts and adds incremental test cases.

The remaining memory leak has to do with contextual EAV nodes that become posteriors due to Hebbian learning. This
may require fancier culling logic to handle since it requires knowledge of when a node activation can affect state.

On the other hand it's probably reasonable to omit these EAV nodes. I suspect these are the nodes resulting from
setting the parent context on a child context, which given the ephemeral nature of the context node kind of defeats
the purpose of having an EAV node dedicated to it.