Switch from ThresholdDistribution to UnimodalHypothesis (and other delusions)
These changes were made with the intention of making negative feedback work as expected for reinforcement debugging.
They did not, but they did improve negative feedback stability to the point where the test still passes often even
with a reasonably high negative feedback on failure. In retrospect it seems reasonable that negative feedback at the
synpatic level may be detrimental for this use case.

These changes also greatly improve k-armed bandit reliability and open the door for node splitting when the
unimodal hypothesis for edges is violated.

Other changes:

* Added a max stack depth and dedicated context hierarchy tracking, which is now also used for recursive
  reinforcement.
* Replaced window-based Hebbian reinforcement with per-synapse outcome-based Hebbian reinforcement. This style is
  more symmetric and stable, and actually replaces stochastic exploration with guided hypothesis testing. This allows
  us to replace ThresholdDistribution with the (for now) non-stochastic UnimodalHypothesis while actually improving
  performance. This can be seen somewhat as an SNN analogue of gradient descent.
* Moves node recency queue promotion from the context to the node contextual state, which allows for more precise
  control over the timing and more streamlined avoidance of the lapsed listener problem.