Begin implementing language bootstrap
Begins by exploring "string literals". This is unlikely to be the ultimate starting point but is an easy place to get
a feel for how this will behave.

The register unbinding story in this iteration is unsatisfactory, and future iterations will use an alternate
sequence recognition and state mechanic.

Also expands binary decoder to 21 bits for unicode code points, and other housekeeping tasks.