Allow GatedBiCluster outputs to be posteriors
This is required to bind context output nodes to STM registers, which is required to support property binding and
object reference.

An alternative may have been to build a latch mechanism directly into GatedBiCluster (effectively a built-in
register), but this is more flexible as it allows any register to bind.