Clean up clusters from LRU
This should make for quicker cleanups as we can stop iterating once we start finding nodes that have been referenced.
The first pass at this logic just stops after the first non-GC'd node it encounters. This should be okay for now,
though longer term it may be susceptible to leaks in the presence of nodes that are referenced but never activated.