Freeze time while busy
Modifies FlexTimeScheduler to only advance time while idle or waiting, allowing for more deterministic timing
especially while debugging.

Also modifies the parse loop to be on a more uniform cadence (relative to loop start rather than loop end). This gets
us to the point of recognizing n-grams, but will need to be modified to support loop blocking for longer operations
if those need to happen in the middle of n-grams (possibly they won't, since they'll be like binding operations).