Consistently suppress posteriors while capturing
This is probably a good idea to prevent unintended side effects.

This is also an example of functionality that is not currently straightforward to factor out into a node due to
timing sensitivity. Furthermore there is some evidence that a more neurologically inspired mechanism would be
inhibition by default and selective disinhibition, but as it is now, that would actually not work as well
timing-wise.